refactor(agent): extract summary prompt module; fix eval judge template
- Move SUMMARY_SYSTEM_PROMPT and SUMMARY_OUTPUT_TOOL into a sibling
  sessionSummaryPrompt.ts so the summarization eval can import them
  without pulling in browser-only modules (authFetch, window.Config).
- Switch the summary-quality judge prompt from {var} to Mustache
  {{var}} syntax so phoenix-evals actually substitutes userMessage,
  assistantMessage, and output before calling the classifier.

Verified end-to-end against a local Phoenix: 10/10 valid format and
10/10 accurate quality on the golden dataset.
