+++ app/src/components/agent/useGenerateSessionSummary.ts
@@ -6,39 +6,17 @@ import {
 } from "ai";
 import { useCallback, useRef } from "react";
 
-import type { FrontendToolDefinition } from "@phoenix/agent/tools/types";
 import { authFetch } from "@phoenix/authFetch";
 import { useAgentStore } from "@phoenix/contexts/AgentContext";
 
 import {
+  SUMMARY_OUTPUT_TOOL,
+  SUMMARY_SYSTEM_PROMPT,
+} from "./sessionSummaryPrompt";
+import {
   getFirstAssistantMessageText,
   getFirstUserMessageText,
 } from "./sessionSummaryUtils";
-
-export const SUMMARY_SYSTEM_PROMPT =
-  "You are a concise summarizer. You MUST call the `summary` tool with a 5-10 word summary of the conversation topic. No quotes, no punctuation at the end.";
-
-/**
- * Output tool that the model is *forced* to call (via `tool_choice: required`)
- * to produce a structured summary. Sent as an `output_tool` rather than a
- * regular `tool` so the server sets `allow_text_output=false` on the
- * pydantic-ai request, preventing the model from responding with free text.
- */
-export const SUMMARY_OUTPUT_TOOL: FrontendToolDefinition = {
-  name: "summary",
-  description: "Provide the conversation summary",
-  parameters: {
-    type: "object",
-    properties: {
-      summary: {
-        type: "string",
-        description: "A 5-10 word summary of the conversation topic",
-      },
-    },
-    required: ["summary"],
-    additionalProperties: false,
-  },
-};
 
 /**
  * Sends a lightweight summarization request through a {@link DefaultChatTransport}
